requirements.txt
addict==2.4.0
aiohttp==3.8.1
aiosignal==1.2.0
alabaster==0.7.12
alembic==1.7.7
altair==4.2.0
anyio==3.4.0
appdata==2.1.2
apptools==5.1.0
apturl==0.5.2
argon2-cffi==21.1.0
asgiref==3.5.0
astor==0.8.1
async-timeout==4.0.2
attrs==19.3.0
autobahn==17.10.1
Automat==0.8.0
autopep8==1.6.0
Babel==2.9.1
backcall==0.2.0
backports.entry-points-selectable==1.1.1
backports.zoneinfo==0.2.1
base58==2.1.1
bcrypt==3.1.7
bleach==4.1.0
blinker==1.4
Brlapi==0.7.0
cachetools==5.0.0
cbor==1.0.0
certifi==2019.11.28
cffi==1.15.0
channels==3.0.4
chardet==3.0.4
charset-normalizer==2.0.7
chrome-gnome-shell==0.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
convert360==0.1.3
cryptography==2.8
cupshelpers==1.0
cupy-cuda114==10.0.0
cycler==0.11.0
Cython==0.29.14
daphne==3.0.2
datify==1.0.4
dbus-python==1.2.16
debugpy==1.5.1
decorator==5.1.0
defer==1.0.6
defusedxml==0.7.1
deprecation==2.1.0
distlib==0.3.3
distro==1.4.0
distro-info===0.23ubuntu1
Django==4.0.3
djlint==1.0.1
dnspython==1.16.0
docopt==0.6.2
docutils==0.17.1
duplicity==0.8.12.0
entrypoints==0.3
envisage==6.0.1
equilib==0.0.1
fasteners==0.14.1
fastrlock==0.8
fett==0.3.2
filelock==3.4.0
Flask==2.1.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
frozenlist==1.2.0
future==0.18.2
gitdb==4.0.9
GitPython==3.1.26
greenlet==1.1.2
html5lib==1.0.1
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
imageio==2.13.0
imagesize==1.2.0
importlib-metadata==4.11.3
importlib-resources==5.4.0
imutils==0.5.4
incremental==16.10.1
ipykernel==6.4.2
ipython==7.28.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.1.2
jedi==0.18.0
Jinja2==3.1.2
joblib==1.1.0
json5==0.9.6
jsonschema==4.1.2
jupyter==1.0.0
jupyter-client==7.0.6
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyter-packaging==0.11.1
jupyter-server==1.12.1
jupyterlab==3.2.4
jupyterlab-pygments==0.1.2
jupyterlab-server==2.8.2
jupyterlab-widgets==1.0.2
keyring==18.0.1
kiwisolver==1.3.2
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
livereload==2.6.3
lockfile==0.12.2
louis==3.12.0
lz4==3.0.2+dfsg
m2r2==0.3.2
macaroonbakery==1.3.1
Mako==1.2.0
MarkupSafe==2.1.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mayavi==4.7.4
MeshLabXML==2018.3
mistune==0.8.4
monotonic==1.5
more-itertools==4.2.0
mpi4py==3.0.3
mpmath==1.2.1
multidict==5.2.0
natsort==7.1.1
nbclassic==0.3.4
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nbsphinx==0.8.7
nest-asyncio==1.5.1
netifaces==0.10.4
networkx==2.6.3
notebook==6.4.5
numpy==1.22.1
oauthlib==3.1.0
olefile==0.46
open3d==0.13.0
opencv-python==4.5.4.58
packaging==21.2
paho-mqtt==1.6.1
pandas==1.3.4
pandocfilters==1.5.0
paramiko==2.6.0
parso==0.8.2
pathspec==0.9.0
pbr==5.8.0
pexpect==4.6.0
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==2.4.0
plyfile==0.7.4
prometheus-client==0.11.0
prompt-toolkit==3.0.21
protobuf==3.6.1
ptyprocess==0.7.0
py-ubjson==0.14.0
py360convert==0.1.0
pyarrow==7.0.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycodestyle==2.8.0
pycparser==2.20
pycups==1.9.73
pydantic==1.9.1
pydeck==0.7.1
pyequilib==0.5.1
pyface==7.3.0
Pygments==2.10.0
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
pymongo==3.11.4
Pympler==1.0.1
PyNaCl==1.3.0
PyOpenGL==3.1.5
pyOpenSSL==19.0.0
pyparsing==2.4.7
pypng==0.0.20
PyQRCode==1.2.1
PyQt5==5.14.1
PyQt5-sip==12.9.0
pyRFC3339==1.1
pyrsistent==0.18.0
PySocks==1.6.8
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-dotenv==0.20.0
python-editor==1.0.4
python-jsonrpc-server==0.3.4
python-snappy==0.5.3
PyTrie==0.2
pytz==2019.3
pytz-deprecation-shim==0.1.0.post0
pyxdg==0.26
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.1.1
QtPy==1.11.2
regex==2022.4.24
reportlab==3.5.34
requests==2.26.0
requests-unixsocket==0.2.0
roman==2.0.0
rstcheck==3.3.1
scikit-learn==1.0.1
scipy==1.7.1
SecretStorage==2.3.1
Send2Trash==1.8.0
service-identity==18.1.0
simplejson==3.16.0
six==1.16.0
sklearn==0.0
smmap==5.0.0
sniffio==1.2.0
snooty-lextudio==1.11.6
snowballstemmer==2.1.0
Sphinx==4.2.0
sphinx-autobuild==2021.3.14
sphinx-rtd-theme==1.0.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==1.4.36
sqlparse==0.4.2
ssh-import-id==5.10
streamlit==1.5.0
sympy==1.9
systemd-python==234
tdqm==0.0.1
terminado==0.12.1
testpath==0.5.0
testresources==2.0.1
threadpoolctl==3.0.0
to-requirements.txt==1.1.8
toml==0.10.2
tomli==2.0.1
tomlkit==0.7.2
toolz==0.11.2
torch==1.10.0
torchaudio==0.10.0
torchsummary==1.5.1
torchvision==0.11.1
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
traits==6.3.2
traitsui==7.2.1
Twisted==18.9.0
txaio==2.10.0
typing-extensions==3.10.0.2
tzdata==2021.5
tzlocal==4.1
u-msgpack-python==2.1
ubuntu-advantage-tools==27.7
ubuntu-drivers-common==0.0.0
ufw==0.36
ujson==1.35
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
validators==0.18.2
virtualenv==20.10.0
vtk==9.1.0
wadllib==1.3.3
watchdog==1.0.2
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.1
Werkzeug==2.1.2
widgetsnbextension==3.5.2
wsaccel==0.6.2
wslink==1.3.1
WTForms==3.0.1
xkit==0.0.0
yarl==1.7.2
zipp==3.7.0
zope.interface==4.7.1
